+++ console/src/dataSources/services/mssql/index.tsx
@@ -197,6 +197,7 @@ export const supportedFeatures: SupportedFeaturesType = {
     retries: false,
     pool_timeout: false,
     connection_lifetime: false,
+    ssl_certificates: false,
   },
 };