+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - console: add `pool_timeout`, `connection_lifetime` and `isolation_level` connection params to connect database form
 - console: add check constraints and comments to MS SQL Server tables' read-only modify page
 - console: add create table functionality for MS SQL Server 
+- console: update connect database form with SSL certificates
 
 ## v2.0.0-alpha.11